Created build.sbt, structured build, preparing for sbt upgrade.
This commit introduces the following new structure:
- build.sbt - the projects
- project/Settings.scala - settings for the projects
- project/Dependencies.scala - external dependencies (and scala versions)
- project/Docs.scala - all things related to generating html and scaladoc docs
- project/FMPP.scala - FreeMarker templating stuff
- project/Osgi.scala - Osgi stuff
- project/TypeProviders.scala - TypeProviders for tests
- project/BuildUtils.scala - misc functions
