+++ project/BuildUtils.scala
@@ -1,0 +1,13 @@
+import sbt._
+
+object BuildUtils {
+  def runTasksSequentially(tasks: List[TaskKey[_]])(state: State): State = tasks match {
+    case t :: ts =>
+      Project.runTask(t.asInstanceOf[TaskKey[Any]], state) match {
+        case None => state.fail
+        case Some((s, Inc(_))) => s.fail
+        case Some((s, _)) => runTasksSequentially(ts)(s)
+      }
+    case Nil => state
+  }
+}
